+++ internal/configs/configtesting/synth_body.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: BUSL-1.1
 
-package configs
+package configtesting
 
 import (
 	"fmt"
